Add support for time periods to tasks, and keep track of state using a state machine
A new time period will be created for a task when it is stopped or paused.

We now have a little state machine we use to keep track of state for a task
(running, stopped, paused).

Improve output of list command - column alignment and time period start and end

Set Time.zone so times will be retrieved from the db in our time zone